added tryResolve to guidReference
TryResolve will attempt to resolve the reference and call OnGuidAdded if the reference can be resolved at the time of the call, without the need to use the gameObject property.

This change is to enable OnGuidAdded as the primary way to cache a refereced component regardless if the guid can be resolved at the time of the call or not.